+++ sqlsynthgen/main.py
@@ -16,8 +16,20 @@ from sqlsynthgen.settings import get_settings
 
 app = typer.Typer()
 
+
 def import_file(file_path: str) -> ModuleType:
-    """Import a file given a relative path."""
+    """Import a file
+
+    This utility function returns
+    the file at file_path as a module
+
+    Args:
+        file_path (str): Path to file to be imported
+
+    Returns:
+        ModuleType
+    """
+
     file_path_path = Path(file_path)
     module_path = ".".join(file_path_path.parts[:-1] + (file_path_path.stem,))
     return import_module(module_path)
@@ -36,7 +48,31 @@ def create_data(
     ssg_file: str = typer.Argument(...),
     num_rows: int = typer.Argument(...),
 ) -> None:
-    """Fill tables with synthetic data."""
+    """Populate schema with synthetic data
+    This CLI command generates synthetic data for
+    Python table structures, and inserts these rows
+    into a destination schema.
+
+    Also takes as input object relational model as represented
+    by file containing Python classes and its attributes.
+
+    Takes as input sqlsynthgen output as represented by Python
+    classes, its attributes and methods for generating values
+    for those attributes.
+
+    Final input is the number of rows required.
+
+    Example:
+        $ python sqlsynthgen/main.py create-data example_orm.py expected_ssg.py 100
+
+    Args:
+        orm_file (str): Path to object relational model.
+        ssg_file (str): Path to sqlsyngen output.
+        num_rows (int): Number of rows of values required
+
+    Returns:
+        None
+    """
     orm_module = import_file(orm_file)
     ssg_module = import_file(ssg_file)
     create_db_data(
@@ -53,15 +89,19 @@ def create_vocab(ssg_file: str = typer.Argument(...)) -> None:
 
 @app.command()
 def create_tables(orm_file: str = typer.Argument(...)) -> None:
-    # """Create tables using the SQLAlchemy file."""
-    """
-    Create tables using the SQLAlchemy file.
+    """Create schema from Python classes
 
-    :param orm_file: Mandatory path to SQLAlchemy file
-    :type kind: str
-    :raise lumache.InvalidKindError: If the kind is invalid.
-    :return: None
-    :rtype: None
+    This CLI command creates Postgresql schema using object relational model
+    declared as Python tables. (eg.)
+
+    Example:
+        $ python sqlsynthgen/main.py create-tables example_orm.py
+
+    Args:
+        orm_file (str): Path to Python tables file.
+
+    Returns:
+        None
 
     """
     orm_module = import_file(orm_file)
@@ -73,7 +113,22 @@ def make_generators(
     orm_file: str = typer.Argument(...),
     config_file: Optional[str] = typer.Argument(None),
 ) -> None:
-    """Make a SQLSynthGen file of generator classes."""
+    """Make a SQLSynthGen file of generator classes.
+
+    This CLI command takes an object relation model output by sqlcodegen and
+    returns a set of synthetic data generators for each attribute
+
+    Example:
+        $ python sqlsynthgen/main.py make-generators example_orm.py
+
+    Args:
+        orm_file (str): Path to Python tables file.
+        config_file (str): Path to configuration file.
+
+    Returns:
+        None
+
+    """
     orm_module = import_file(orm_file)
     generator_config = read_yaml_file(config_file) if config_file is not None else {}
     result = make_generators_from_tables(orm_module, generator_config)
@@ -82,7 +137,22 @@ def make_generators(
 
 @app.command()
 def make_tables() -> None:
-    """Make a SQLAlchemy file of Table classes."""
+    """Make a SQLAlchemy file of Table classes.
+
+    This CLI command deploys sqlacodegen to discover a
+    schema structure, and generates a object relational model declared
+    as Python classes.
+
+    Example:
+        $ python sqlsynthgen/main.py make_tables
+
+    Args:
+        None
+
+    Returns:
+        None
+
+    """
     settings = get_settings()
 
     command = ["sqlacodegen"]
@@ -101,6 +171,7 @@ def make_tables() -> None:
         sys.exit(e.returncode)
 
     print(completed_process.stdout)
-    
+
+
 if __name__ == "__main__":
     app()